fix: move xxhash reflection utility
