+++ ZeepSDK.Tests/ZeepSDK.Tests.csproj
@@ -25,11 +25,11 @@
         <Compile Include="..\ZeepSDK\Level\CsvZeepLevelParser.cs" Link="Source\CsvZeepLevelParser.cs" />
         <Compile Include="..\ZeepSDK\Level\LevelHashV2.cs" Link="Source\LevelHashV2.cs" />
         <Compile Include="..\ZeepSDK\Level\LevelHashV2Calculator.cs" Link="Source\LevelHashV2Calculator.cs" />
-        <Compile Include="..\ZeepSDK\Level\XxHash128Reflection.cs" Link="Source\XxHash128Reflection.cs" />
         <Compile Include="..\ZeepSDK\Numerics\Vector2.cs" Link="Source\Vector2.cs" />
         <Compile Include="..\ZeepSDK\Numerics\Vector3.cs" Link="Source\Vector3.cs" />
         <Compile Include="..\ZeepSDK\Utilities\SequenceComparer.cs" Link="Source\SequenceComparer.cs" />
         <Compile Include="..\ZeepSDK\Utilities\Vector3Comparer.cs" Link="Source\Vector3Comparer.cs" />
+        <Compile Include="..\ZeepSDK\Utilities\XxHash128Reflection.cs" Link="Source\XxHash128Reflection.cs" />
     </ItemGroup>
 
     <ItemGroup>
@@ -38,10 +38,3 @@
         </None>
     </ItemGroup>
 </Project>
-
-
-
-
-
-
-
