+++ ZeepSDK/Level/LevelHashV2Calculator.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Text;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using ZeepSDK.Utilities;
 
 namespace ZeepSDK.Level;
